classification-service/requirements.txt
requests~=2.32.3
torch~=2.5.1
Flask~=3.1.0
torchvision~=0.20.1
pillow~=11.0.0
matplotlib~=3.10.0
python-consul~=1.1.0